+++ src/lightning_utilities/docs/__init__.py
@@ -1,3 +1,6 @@
 """General tools for Docs."""
 
-from lightning_utilities.docs.retriever import fetch_external_assets  # noqa: F401
+from lightning_utilities.docs.formatting import adjust_linked_external_docs
+from lightning_utilities.docs.retriever import fetch_external_assets
+
+__all__ = ["adjust_linked_external_docs", "fetch_external_assets"]